src/parsers/filing_parser.py
"""
Filing parser for extracting text, tables, and figures from 10-K documents
"""
from pathlib import Path
from typing import Dict, List, Any, Optional
import re


class FilingParser:
    """Parse 10-K filings (PDF or HTML) into structured components"""

    def __init__(self):
        """Initialize parser with necessary libraries"""
        # In practice, you would import:
        # - docling or PyMuPDF for PDF parsing
        # - BeautifulSoup for HTML parsing
        # - Tesseract for OCR
        pass

    def parse(
        self,
        file_path: str,
        file_format: str,
        ticker: str,
        fiscal_year: int
    ) -> Dict[str, Any]:
        """
        Parse a 10-K filing

        Args:
            file_path: Path to the filing document
            file_format: Format (html or pdf)
            ticker: Company ticker
            fiscal_year: Fiscal year

        Returns:
            Parsed document dictionary with text, tables, figures
        """
        if file_format == 'html':
            return self._parse_html(file_path, ticker, fiscal_year)
        elif file_format == 'pdf':
            return self._parse_pdf(file_path, ticker, fiscal_year)
        else:
            raise ValueError(f"Unsupported format: {file_format}")

    def _parse_html(self, file_path: str, ticker: str, fiscal_year: int) -> Dict[str, Any]:
        """Parse HTML filing"""
        from bs4 import BeautifulSoup

        try:
            with open(file_path, 'r', encoding='utf-8', errors='ignore') as f:
                html_content = f.read()

            soup = BeautifulSoup(html_content, 'html.parser')

            # Remove script and style elements
            for script in soup(["script", "style"]):
                script.decompose()

            # Extract all text
            text = soup.get_text(separator=' ', strip=True)

            # Clean up whitespace
            text = re.sub(r'\s+', ' ', text)

            # Split into sections (simple chunking approach)
            sections = []
            chunk_size = 5000
            for i in range(0, min(len(text), 50000), chunk_size):  # Limit to first 50k chars
                if i + chunk_size > len(text):
                    break
                sections.append({
                    'title': f'Section {i // chunk_size + 1}',
                    'content': text[i:i + chunk_size]
                })

            # Extract tables (basic - find <table> tags)
            tables = []
            html_tables = soup.find_all('table')
            for idx, table in enumerate(html_tables[:10]):  # Limit to first 10 tables
                tables.append({
                    'table_id': f'T{idx+1}',
                    'html': str(table)
                })

            return {
                'ticker': ticker,
                'fiscal_year': fiscal_year,
                'format': 'html',
                'num_pages': None,
                'text': sections,
                'tables': tables,
                'figures': []
            }

        except Exception as e:
            print(f"Error parsing HTML for {ticker} {fiscal_year}: {str(e)}")
            return {
                'ticker': ticker,
                'fiscal_year': fiscal_year,
                'format': 'html',
                'num_pages': None,
                'text': [],
                'tables': [],
                'figures': []
            }

    def _parse_pdf(self, file_path: str, ticker: str, fiscal_year: int) -> Dict[str, Any]:
        """Parse PDF filing"""
        # Implementation using docling/PyMuPDF
        # This is a placeholder - actual implementation would:
        # 1. Load PDF with structure preservation
        # 2. Extract text with layout information
        # 3. Identify and parse tables with headers
        # 4. OCR figures and extract captions
        # 5. Track page numbers and bounding boxes

        return {
            'ticker': ticker,
            'fiscal_year': fiscal_year,
            'format': 'pdf',
            'num_pages': 0,  # Actual page count
            'text': [],  # List of text sections with page numbers
            'tables': [],  # List of table objects with structure
            'figures': []  # List of figures with captions
        }

    def extract_metadata(self, parsed_doc: Dict) -> Dict[str, Any]:
        """Extract metadata from parsed document"""
        metadata = {
            'ticker': parsed_doc['ticker'],
            'fiscal_year': parsed_doc['fiscal_year'],
            'format': parsed_doc['format'],
            'num_pages': parsed_doc.get('num_pages'),
            'num_tables': len(parsed_doc.get('tables', [])),
            'num_figures': len(parsed_doc.get('figures', []))
        }
        return metadata
