requirements.txt
# Multi-Modal RAG for SEC 10-K Filings
# Requirements for CS582 Graduate Project

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Document parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyMuPDF>=1.23.0  # PDF parsing
python-docx>=0.8.11

# NLP and embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0  # or tensorflow
tokenizers>=0.15.0

# Retrieval and indexing
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
rank-bm25>=0.2.2

# LLM inference (choose based on setup)
openai>=1.3.0  # For OpenAI API
# OR for local models:
# llama-cpp-python>=0.2.0
# ctransformers>=0.2.0

# Web scraping and APIs
requests>=2.31.0
urllib3>=2.0.0

# Data handling
tqdm>=4.66.0
python-dotenv>=1.0.0

# Evaluation
scikit-learn>=1.3.0
nltk>=3.8.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# OCR (optional, for figures)
pytesseract>=0.3.10
pillow>=10.0.0

# XBRL parsing (optional)
# arelle>=2.0  # For XBRL data

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Optional: Cross-encoder for reranking
# sentence-transformers already includes this

# System
pathlib>=1.0.1
python-dateutil>=2.8.0
